Git Ignore - removed cached tracking
I added some stuff to the git ignore but it was still being tracked. Apparently once it tracks a file, it keeps a reference to it in it's index and only uses the git ignore for newly appearing files/locations. So since we started the project before adding a git ignore it had already indexed /Library/. I used the command "git rm -r --cached /Library/" to remove it from the index and now it will stop tracking anything in that folder. A handy approach if we need it but I'll still read up on standard practices in the next couple days.